Update update-azure-cli.md
we suggest the gray highlight for referring to code/command. here we don't have code named 'update', I believe you are referring to the sections in the article? manual/automatic? in that case bold is appropriate to catch attention. but if you were referring to code, like you do with 'az upgrade' then gray highlight makes sense. thanks for applying these to make the text visually distinct :)